refactor(sku): split AdminSkuMapeamento (503→100) em hook + componentes
Extrai verbatim de src/pages/AdminSkuMapeamento.tsx:
- useSkuMapeamento: 2 queries (mapeamentos + descrições), memos de filtro
  (empresas/fornecedores/filtrados), upsert e validação (faltantes/suspeitos)
- MapeamentoFiltros, MapeamentosTable, MapeamentoFormDialog, ValidacaoDialog,
  types.ts, config.ts (EMPTY_FORM/SkuMapForm)
- +14 testes (presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
